cheri: Clean up the purecap UABI
The purecap UABI uses __kernel_uintptr_t in UABI headers if upstream
uses an integer that can contain a pointer. This is a problem because
some UABI headers on non-cheri systems use "unsigned long" for while
others use __u64. Replacing both of these with __kernel_uintptr_t will
change the non-cheri UABI for one of these types on 32-bit systems
where unsigned long is 32-bit bit __u64 is 64-bit.

Cleanup the UABI header files and define new types for both cases:
- __u64ptr is always at least 64-bit and can hold a pointer on CHERI
  systems. Use this to replace __u64. There is also an aligned variant
  __aligned_u64ptr that replaces __aligned_u64.
- __uptr has the same size as an unsigned long on non-cheri systems.
  Use this for pointers that are stored in unsigned long fields.

Resurrect u64_to_user_ptr but require that at __u64ptr is passed to it.
This handles problems on 32-bit non-cheri systems where a direct cast
between __u64ptr (which is 64-bit) and void * (which is 32-bit) triggers
a compiler error.

Convert remaining plain addresses in UABI headers to pointers. This
removes almost all uses of uaddr_to_user_ptr().

Here's a list of the new UABI types an their sizes for reference:

TYPE		64-bit		64-bit		32-bit		32-bit
		cheri		non-cheri	cheri		non-cheri

__uptr		128		64		64		32
unsigned long	 64		64		32		32
__u64ptr	128		64		64		64
__u64		 64		64		64		64

+++ sound/pci/hda/hda_local.h
@@ -656,15 +656,15 @@ void snd_hda_codec_shutdown(struct hda_codec *codec);
  * AMP control callbacks
  */
 /* retrieve parameters from private_value */
-#define get_amp_nid_(pv)	((pv) & 0xffff)
+#define get_amp_nid_(pv)	((unsigned long)(pv) & 0xffff)
 #define get_amp_nid(kc)		get_amp_nid_(__c_ua((kc)->private_value))
-#define get_amp_channels(kc)	((__c_ua((kc)->private_value) >> 16) & 0x3)
-#define get_amp_direction_(pv)	(((pv) >> 18) & 0x1)
+#define get_amp_channels(kc)	((unsigned long)(__c_ua((kc)->private_value) >> 16) & 0x3)
+#define get_amp_direction_(pv)	(((unsigned long)(pv) >> 18) & 0x1)
 #define get_amp_direction(kc)	get_amp_direction_(__c_ua((kc)->private_value))
-#define get_amp_index_(pv)	(((pv) >> 19) & 0xf)
+#define get_amp_index_(pv)	(((unsigned long)(pv) >> 19) & 0xf)
 #define get_amp_index(kc)	get_amp_index_(__c_ua((kc)->private_value))
-#define get_amp_offset(kc)	((__c_ua((kc)->private_value) >> 23) & 0x3f)
-#define get_amp_min_mute(kc)	((__c_ua((kc)->private_value) >> 29) & 0x1)
+#define get_amp_offset(kc)	((unsigned long)(__c_ua((kc)->private_value) >> 23) & 0x3f)
+#define get_amp_min_mute(kc)	((unsigned long)(__c_ua((kc)->private_value) >> 29) & 0x1)
 
 /*
  * enum control helper
